Move API to api.rs + global_fns.rs
